Initialise TIST with consistent system time at startup
The change introduced in fd33dd4 (v2.4.0) had the consequence of
generating different timestamp offsets depending on when exactly
the mux was started. The intended goal to avoid having the NULL
symbol output at a fixed offset to avoid receiver synchronisation is
however not so important, because most commercial receivers will
cut transmission for several seconds before resuming, and the majority
of the receivers will then start resynchronising.

However, the different timestamp offsets upsets some modulator input
modules, which so not necessarily have enough buffering capacity.

This change reverts back to the previous behaviour, where the timestamp
is set to current system time, rounded to 24ms.