Added FIG 0/14 for FEC scheme defintion, when a packet model channel is using FEC error correction (enhancedpacket)
Needs another pair of eyes to check it.

FIG0_14.cpp, line 37 - the data field is a single byte, top 6 bits are SubChanID, bottom 2 bits are FEC type; is this how you organise the bits in the right order?
FIG0_14.cpp, line 78 - there needs to be a test here if the packet channel is enhanced or not, but it looks like that attribute is only held on the file input, and not easily readable from the services configuration?
FIG0_14.cpp, lines 101-106 - is this writing the entire FIC into the stack correctly? The usual preamble bytes for the service identifier followed by the 1 byte data field